LayoutTests/imported/w3c/web-platform-tests/css/css-values/sin-cos-tan-serialize-expected.txt


FAIL 'cos(0)' as a specified value should serialize as 'calc(1)'. assert_equals: 'cos(0)' and 'calc(1)' should serialize the same in specified values. expected "calc(1)" but got "cos(0)"
FAIL 'scale(cos(0))' as a specified value should serialize as 'scale(calc(1))'. assert_equals: 'scale(cos(0))' and 'scale(calc(1))' should serialize the same in specified values. expected "scale(calc(1))" but got "scale(cos(0))"
PASS 'cos(0)' as a computed value should serialize as '1'.
PASS 'scale(cos(0))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'.
FAIL 'sin(0)' as a specified value should serialize as 'calc(0)'. assert_equals: 'sin(0)' and 'calc(0)' should serialize the same in specified values. expected "calc(0)" but got "sin(0)"
FAIL 'scale(sin(0))' as a specified value should serialize as 'scale(calc(0))'. assert_equals: 'scale(sin(0))' and 'scale(calc(0))' should serialize the same in specified values. expected "scale(calc(0))" but got "scale(sin(0))"
PASS 'sin(0)' as a computed value should serialize as '0'.
PASS 'scale(sin(0))' as a computed value should serialize as 'matrix(0, 0, 0, 0, 0, 0)'.
FAIL 'tan(0)' as a specified value should serialize as 'calc(0)'. assert_equals: 'tan(0)' and 'calc(0)' should serialize the same in specified values. expected "calc(0)" but got "tan(0)"
FAIL 'scale(tan(0))' as a specified value should serialize as 'scale(calc(0))'. assert_equals: 'scale(tan(0))' and 'scale(calc(0))' should serialize the same in specified values. expected "scale(calc(0))" but got "scale(tan(0))"
PASS 'tan(0)' as a computed value should serialize as '0'.
PASS 'scale(tan(0))' as a computed value should serialize as 'matrix(0, 0, 0, 0, 0, 0)'.
PASS 'calc(sin(0) + cos(0) + tan(0))' as a specified value should serialize as 'calc(1)'.
PASS 'scale(calc(sin(0) + cos(0) + tan(0)))' as a specified value should serialize as 'scale(calc(1))'.
PASS 'calc(sin(0) + cos(0) + tan(0))' as a computed value should serialize as '1'.
PASS 'scale(calc(sin(0) + cos(0) + tan(0)))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'.
PASS 'calc(sin(0) + 0.5)' as a specified value should serialize as 'calc(0.5)'.
PASS 'scale(calc(sin(0) + 0.5))' as a specified value should serialize as 'scale(calc(0.5))'.
PASS 'calc(sin(0) + 0.5)' as a computed value should serialize as '0.5'.
PASS 'scale(calc(sin(0) + 0.5))' as a computed value should serialize as 'matrix(0.5, 0, 0, 0.5, 0, 0)'.
PASS 'calc(cos(0) + 0.5)' as a specified value should serialize as 'calc(1.5)'.
PASS 'scale(calc(cos(0) + 0.5))' as a specified value should serialize as 'scale(calc(1.5))'.
FAIL 'calc(cos(0) + 0.5)' as a computed value should serialize as '1.5'. assert_equals: '1.5' should round-trip exactly in computed values. expected "1.5" but got "1"
PASS 'scale(calc(cos(0) + 0.5))' as a computed value should serialize as 'matrix(1.5, 0, 0, 1.5, 0, 0)'.
PASS 'calc(tan(0) + 0.5)' as a specified value should serialize as 'calc(0.5)'.
PASS 'scale(calc(tan(0) + 0.5))' as a specified value should serialize as 'scale(calc(0.5))'.
PASS 'calc(tan(0) + 0.5)' as a computed value should serialize as '0.5'.
PASS 'scale(calc(tan(0) + 0.5))' as a computed value should serialize as 'matrix(0.5, 0, 0, 0.5, 0, 0)'.
PASS 'calc(sin(infinity))' as a specified value should serialize as 'calc(NaN)'.
PASS 'scale(calc(sin(infinity)))' as a specified value should serialize as 'scale(calc(NaN))'.
FAIL 'calc(sin(infinity))' as a computed value should serialize as 'NaN'. assert_equals: 'NaN' should round-trip exactly in computed values. expected "NaN" but got "1"
FAIL 'scale(calc(sin(infinity)))' as a computed value should serialize as 'matrix(NaN, 0, 0, NaN, 0, 0)'. assert_equals: 'matrix(NaN, 0, 0, NaN, 0, 0)' should round-trip exactly in computed values. expected "matrix(NaN, 0, 0, NaN, 0, 0)" but got "none"
PASS 'calc(cos(infinity))' as a specified value should serialize as 'calc(NaN)'.
PASS 'scale(calc(cos(infinity)))' as a specified value should serialize as 'scale(calc(NaN))'.
FAIL 'calc(cos(infinity))' as a computed value should serialize as 'NaN'. assert_equals: 'NaN' should round-trip exactly in computed values. expected "NaN" but got "1"
FAIL 'scale(calc(cos(infinity)))' as a computed value should serialize as 'matrix(NaN, 0, 0, NaN, 0, 0)'. assert_equals: 'matrix(NaN, 0, 0, NaN, 0, 0)' should round-trip exactly in computed values. expected "matrix(NaN, 0, 0, NaN, 0, 0)" but got "none"
PASS 'calc(tan(infinity))' as a specified value should serialize as 'calc(NaN)'.
PASS 'scale(calc(tan(infinity)))' as a specified value should serialize as 'scale(calc(NaN))'.
FAIL 'calc(tan(infinity))' as a computed value should serialize as 'NaN'. assert_equals: 'NaN' should round-trip exactly in computed values. expected "NaN" but got "1"
FAIL 'scale(calc(tan(infinity)))' as a computed value should serialize as 'matrix(NaN, 0, 0, NaN, 0, 0)'. assert_equals: 'matrix(NaN, 0, 0, NaN, 0, 0)' should round-trip exactly in computed values. expected "matrix(NaN, 0, 0, NaN, 0, 0)" but got "none"
